api/requirements.txt
Django==3.1.7
django-compat==1.0.15
djangorestframework==3.12.2
psycopg2==2.8.6
uWSGI==2.0.19.1
